Move dcos_installer.util to dcos_installer.constants, it only has constants
