+++ ext/dcos-installer/dcos_installer/config/__init__.py
@@ -13,7 +13,7 @@ import logging
 import os
 import yaml
 
-from dcos_installer.util import CONFIG_PATH, SSH_KEY_PATH, IP_DETECT_PATH
+from dcos_installer.constants import CONFIG_PATH, SSH_KEY_PATH, IP_DETECT_PATH
 log = logging.getLogger(__name__)
