+++ ext/dcos-installer/dcos_installer/constants.py
